Authorization

+++ WebApp/Services/RestaurantService.cs
@@ -1,5 +1,8 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+using WebApp.Authorization;
 using WebApp.Entities;
 using WebApp.Exceptions;
 using WebApp.Models;
@@ -8,14 +11,19 @@ namespace WebApp.Services
 {
     public class RestaurantService : IRestaurantService
     {
+        private readonly IUserContextService _userContextService;
+        private readonly IAuthorizationService _authorizationService;
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
 
         public ILogger<RestaurantService> Logger { get; }
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger
+            , IAuthorizationService authorizationService, IUserContextService userContextService)
         {
+            _userContextService = userContextService;
+            _authorizationService = authorizationService;
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
@@ -45,6 +53,7 @@ namespace WebApp.Services
         public int Create(CreateRestaurantDto dto) 
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
@@ -60,6 +69,14 @@ namespace WebApp.Services
             if(restaurant is null) 
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult =
+                _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
@@ -72,6 +89,14 @@ namespace WebApp.Services
 
             if (restaurant is null)  
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = 
+                _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Descripton = dto.Descripton;
